StyleCop adjustments
